+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/AbstractModel.java
@@ -1156,25 +1156,6 @@ public abstract class AbstractModel {
     }
 
     /**
-     * Generate a Map with Prometheus annotations
-     *
-     * @return Map with Prometheus annotations using the default port (9404) and path (/metrics)
-     */
-    protected Map<String, String> prometheusAnnotations()    {
-        if (isMetricsEnabled) {
-            Map<String, String> annotations = new HashMap<>(3);
-
-            annotations.put("prometheus.io/port", String.valueOf(METRICS_PORT));
-            annotations.put("prometheus.io/scrape", "true");
-            annotations.put("prometheus.io/path", METRICS_PATH);
-
-            return annotations;
-        } else {
-            return null;
-        }
-    }
-
-    /**
      * Creates the PodDisruptionBudget
      *
      * @return The default PodDisruptionBudget